Make list items starting with plus fragments
This introduces some hacky Liquid code to get around the limitations
of the filters available. It’s not possible to use regular expressions
in replacements, so only replacing `+` at the start of a line requires
looping over each line inspecting a substring of the line on each
iteration.

In addition it is not easy to iterate over each line. You have to first
convert newlines to breaks, then split by those, then add the newlines
back in to avoid impacting the formatting. This is considerable extra
complexity but the trade-off is neater markup in the slides at the
expense of less functional Liquid.

I’ve removed the `-[+]` notation in favour of this simpler version.