Fixes #31: indented items render as fragments
This fixes an issue where if you indented an item and used the `+` sign
to make it a fragment, it would simply appear when it’s parent item was
shown. I don’t know why I was using `replace_first` and then `prepend`
as it seems simpler to just use `replace_first` as I have now. This
preserves the indent whitespace of the list item and seems to work
properly.