Add a mechanism to escape '+' fragments
Prior to this change, it was not possible to start a line with `+` at
all, which is particularly problematic when you have a diff in literal
blocks.

Possibly a better long-term fix for this problem would be to disable all
pre-processing between triple-backticks, but the ability to escape `+`
at the beginning of a line seems useful in itself and is much easier to
implement.