Fix issue with non clickable links
The use of HTML elements was wrong, causing reveal.js to think this was a vertical stack, applying a CSS rule to the `<section>` to disable pointer events.

I’ve fixed this, and now use [slide attributes](https://revealjs.com/markdown/#slide-attributes) to add the correct class to the slide so that the mermaid plugin code can detect the diagram.