Provide a way to control diagram height
This is possible using element attributes to set the style on the div, but it feels a little cumbersome and easy to forget for the slide author. I can’t come up with a nicer way to do this using HTML like tags due to the limitations of liquid and the importance of the position of the element attribute.

I’ve also improved the code which makes sure that the SVG fits within the available space in the div on the page as it wasn’t working. I’ve stuck with the same idea of scaling the SVG. This felt like a better solution than resizing the div to fit the content.

I found I had to override the `max-width` of the svg element as the value already there was disrupting the horizontal centering of the diagram. It is necessary to position absolute within the relative wrapper to scale and centre the SVG.

I found that it is better to use `svgElem.getBoundingClientRect().width;` rather than `svgElem.getAttribute('width');` as this provides the actual size of the SVG content.

I’ve ensured that the diagram wouldn’t normally fit in the space available to ensure the scaling code works.