Rerender diagram after overview navigation
This resolves an issue where the diagram is rendered as the user navigates in the overview, then selects a slide, and finds the text is not visible.

The diagram was being rendered while the slide was tiny in the overview, and this meant the sizes were screwed up. By destroying and showing the diagram again when the overview is hidden (hence the slide is the proper size), the diagram renders properly.