Stop marked processing diagram code as Markdown
This was causing issues with Mermaid state diagrams, as the `*` in the start and end nodes (`[*]`) was being interpreted as emphasis (`<em>` tags were finding their way into the DOM). I found inspiration in [this solution](https://revealjs.com/math/#markdown) in the reveal.js docs which uses a single backtick to wrap text in a `<code>` block, which prevents marked (the JS Markdown processor used by reveal.js) from processing the text as markdown. To accommodate this change I’ve tweaked our JS to strip the code tag out of the Mermaid diagram code.

This change may also provide an alternative approach for #35.

I’ve upgraded Mermaid to default to 11.0.2, and users can now override this version in their `config.yml` like this: `mermaid_version: 11.0.2`.

11.0.2 is the most recent which didn’t introduce weird rendering issues with state diagrams which I’ve not spent time investigating.